requirements.txt
accelerate<=1.1.0
addict
datasets
bitsandbytes; sys_platform != 'darwin'
einops
facexlib
future
icecream
lmdb
numpy
openai-clip
opencv-python-headless
pandas
Pillow
pre-commit
pytest
pyyaml
requests
ruff
scikit-image
scipy
sentencepiece
tensorboard
timm>=0.8
torch>=1.12
torchvision>=0.13
tqdm
transformers==4.37.2
yapf
gdown
